+++ src/Ext/Scenario/Body.cpp
@@ -168,6 +168,8 @@ void ScenarioExt::ExtData::Serialize(T& Stm)
 		.Process(this->DefaultLS800BkgdName)
 		.Process(this->DefaultLS800BkgdPal)
 		.Process(this->MasterDetonationBullet)
+		.Process(this->LimboLaunchers)
+//		.Process(this->NewMessageList); // Should not S/L
 		;
 }