+++ src/Ext/Techno/Hooks.Firing.cpp
@@ -436,11 +436,24 @@ DEFINE_HOOK(0x6FDD7D, TechnoClass_FireAt_UpdateWeaponType, 0x5)
 	enum { CanNotFire = 0x6FDE03 };
 
 	GET(TechnoClass* const, pThis, ESI);
+	GET(WeaponTypeClass* const, pWeapon, EBX);
+
+	if (pWeapon->LimboLaunch)
+	{
+		const auto pWH = pWeapon->Warhead;
+
+		if (!pWH->Parasite && WarheadTypeExt::ExtMap.Find(pWH)->UnlimboDetonate)
+		{
+			if (const auto pFoot = abstract_cast<FootClass*, true>(pThis))
+			{
+				if (pFoot->Locomotor->Is_Really_Moving_Now())
+					return CanNotFire;
+			}
+		}
+	}
 
 	if (const auto pExt = TechnoExt::ExtMap.TryFind(pThis))
 	{
-		GET(WeaponTypeClass* const, pWeapon, EBX);
-
 		if (pThis->CurrentBurstIndex && pWeapon != pExt->LastWeaponType && pExt->TypeExtData->RecountBurst.Get(RulesExt::Global()->RecountBurst))
 		{
 			if (pExt->LastWeaponType && pExt->LastWeaponType->Burst)