+++ docs/Whats-New.md
@@ -446,6 +446,8 @@ New:
 - Randomized anims for several behaviors (by Ollerus)
 - Shield respawn animation and weapon (by Ollerus)
 - [Customize the chained damage of the wall](Fixed-or-Improved-Logics.md#customize-the-chained-damage-of-the-wall) (by TaranDahl)
+- Allow the aircraft to enter area guard mission and not crash immediately without any airport (by CrimRecya)
+- Unlimbo Detonate warhead (by FlyStar)
 
 Vanilla fixes:
 - Fixed sidebar not updating queued unit numbers when adding or removing units when the production is on hold (by CrimRecya)
@@ -483,6 +485,8 @@ Phobos fixes:
 - Fixed `SkirmishUnlimitedColors` not being checked if Phobos runs without Ares active (by Starkku)
 - Fixed number of `*.ApplyFirepowerMult` options (f.ex anim damage, crit) ignoring veterancy firepower modifier (by Starkku)
 - Fixed map trigger action `125 Build At...` not always playing buildups correctly (by Starkku)
+- Fixed an issue that jumpjet vehicles can not stop correctly when assigned a target in range. (by TaranDahl)
+- Fixed an issue that jumpjet infantries stop incorrectly when assigned a target out of range. (by TaranDahl)
 
 Fixes / interactions with other extensions:
 - Allowed `AuxBuilding` and Ares' `SW.Aux/NegBuildings` to count building upgrades (by Ollerus)